+++ apps/imadm/src/adm_passport_logic.erl
@@ -15,9 +15,7 @@
 -include_lib("imlib/include/common.hrl").
 
 
-% login_success_transfer 按顺序匹配了
-% 密码校验取第4位为密码数据 lists:nth(5, Row)
--define (LOGIN_COLUMN, <<"id,account,mobile,password,nickname,avatar,role_id,email">>).
+-define (LOGIN_COLUMN, <<"id,account,mobile,password,nickname,avatar,role_id,email,status">>).
 
 %% ===================================================================
 %% API
@@ -30,31 +28,14 @@
 do_login(_Email, <<>>) ->
     {error, "密码有误"};
 do_login(Mobile, Pwd) ->
-    Res =
-        case imboy_func:is_mobile(Mobile) of
-            true ->
-                adm_user_repo:find_by_mobile(Mobile, ?LOGIN_COLUMN);
-            false ->
-                adm_user_repo:find_by_account(Mobile, ?LOGIN_COLUMN)
-        end,
-    % ?LOG(Res),
-    {Check, User} =
-        case Res of
-            {ok, _, [Row]} when is_tuple(Row) ->
-                % ?LOG(['Pwd', Pwd]),
-                % 第四个元素为password
-                case imboy_password:verify(Pwd, element(4, Row)) of
-                    {ok, _} ->
-                        {true, Row};
-                    {error, Msg} ->
-                        {false, Msg}
-                end;
-            _ ->
-                % io:format("res is ~p~n", [Res]),
-                {false, []}
-        end,
-    % ?LOG(['Check', Check, "user" , User]),
-    login_success_transfer(Check, User).
+    User = case imboy_func:is_mobile(Mobile) of
+        true ->
+            adm_user_repo:find_by_mobile(Mobile, ?LOGIN_COLUMN);
+        false ->
+            adm_user_repo:find_by_account(Mobile, ?LOGIN_COLUMN)
+    end,
+    % ?LOG(User),
+    do_login_transfer(Pwd, User).
 
 
 
@@ -62,23 +43,34 @@ do_login(Mobile, Pwd) ->
 %% Internal Function Definitions
 %% ===================================================================-
 
-% <<"id,account,mobile,password,nickname,avatar,role_id,email">>
--spec login_success_transfer(boolean(), tuple()) -> {ok, map()} | {error, any()}.
-login_success_transfer(true, {Id, Account, Mobile, _, Nickname, Avatar, RoleId, Email}) ->
-    {ok, #{
-           <<"id">> => imboy_hashids:encode(Id),
-           <<"mobile">> => Mobile,
-           <<"email">> => Email,
-           <<"nickname">> => Nickname,
-           <<"avatar">> => Avatar,
-           <<"account">> => Account,
-           <<"role_id">> => RoleId
-          }};
-% login_success_transfer(_, User) ->
-%     ?LOG([User]),
-%     {error, "账号或密码错误"}.
-login_success_transfer(_, _) ->
-    {error, "账号或密码错误"}.
+
+-spec do_login_transfer(binary(), map()) -> {ok, map()} | {error, any()}.
+do_login_transfer(Pwd, User) ->
+    Pwd2 = maps:get(<<"password">>, User, <<>>),
+    % 状态: -1 删除  0 禁用  1 启用
+    Status = maps:get(<<"status">>, User, -2),
+    case imboy_password:verify(Pwd, Pwd2) of
+        {ok, _} when Status == -2 ->
+            {error, "账号不存在"};
+        {ok, _} when Status == -1 ->
+            {error, "账号不存在或者已删除"};
+        {ok, _} when Status == 0 ->
+            {error, "账号被禁用"};
+        {ok, _} when Status == 1 ->
+            Id = maps:get(<<"id">>, User),
+            {ok, #{
+               <<"id">> => imboy_hashids:encode(Id),
+               <<"mobile">> => maps:get(<<"mobile">>, User),
+               <<"email">> => maps:get(<<"email">>, User),
+               <<"nickname">> => maps:get(<<"nickname">>, User),
+               <<"avatar">> => maps:get(<<"avatar">>, User),
+               <<"account">> => maps:get(<<"account">>, User),
+               <<"role_id">> => maps:get(<<"role_id">>, User)
+              }};
+        {error, Msg} ->
+            {error, Msg}
+    end.
+
 
 %% ===================================================================
 %% EUnit tests.